Restructured derive macro
